Move setting classes to their own models
