+++ src/models/cloudSettings.model.ts
@@ -1,0 +1,6 @@
+import { Environment } from "../environmentPath";
+
+export class CloudSettings {
+  public lastUpload: Date = null;
+  public extensionVersion: string = "v" + Environment.getVersion();
+}
