+++ src/models/extensionConfig.model.ts
@@ -1,0 +1,9 @@
+export class ExtensionConfig {
+  public gist = null;
+  public quietSync = false;
+  public removeExtensions = true;
+  public syncExtensions = true;
+  public autoDownload = false;
+  public autoUpload = false;
+  public forceDownload = false;
+}
